server/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
flask-cors==6.0.1
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
s3fs==2025.10.0
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.22.0
python-dotenv==1.0.0
requests==2.32.3
google-auth==2.42.1
google-genai==1.47.0

google-genai==1.47.0